refactor: streamline command registration and remove hardcoded commands
