+++ service/bot/embeds/colors.go
@@ -1,0 +1,9 @@
+// Package embeds provides Discord embed creation utilities
+package embeds
+
+// Color constants for message embeds
+const (
+	ColorPrimary = 0x5c5fea // Primary color for standard messages
+	ColorDanger  = 0xd43535 // Danger color for error messages
+	ColorWoW     = 0x00AEEF // World of Warcraft specific color
+)
